pyproject.toml
[tool.poetry]
name = "nonebot_plugin_r6s"
version = "0.4.2"
description = "A R6s plugin for Nonebot2"
authors = ["abrahumlink <307887491@qq.com>"]
documentation = "https://github.com/abrahum/nonebot_plugin_r6s#readme"
license = "GPLv3"
homepage = "https://github.com/abrahum/nonebot_plugin_r6s"
readme = "README.md"
keywords = ["nonebot", "nonebot2", "r6s"]

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "^2.0.0-beta.1"
ujson = ">=4.0.2,<6.0.0"
httpx = ">=0.21.1, <1.0.0"
Pillow = ">=8.4,<10.0"

[tool.poetry.dev-dependencies]


[tool.pdm]
[tool.pdm.dev-dependencies]
dev = []

[tool.pdm.build]
includes = []
[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"


[project]
name = "nonebot_plugin_r6s"
version = "0.4.2"
description = "A R6s plugin for Nonebot2"
authors = [
    {name = "BalconyJH", email = "balconyjh@gmail.com"},
    {name = "abrahumlink",email = "307887491@qq.com"},
]
dependencies = [
    "nonebot2[fastapi]>=2.0.0b5",
    "nonebot-adapter-onebot<3.0.0.0,>=2.0.0-beta.1",
    "ujson<6.0.0,>=4.0.2",
    "httpx<1.0.0,>=0.21.1",
    "Pillow<10.0,>=8.4",
]
requires-python = ">=3.7.3,<4.0"
license = {text = "GPLv3"}
readme = "README.md"
keywords = ["nonebot", "nonebot2", "r6s"]

[project.urls]
homepage = "https://github.com/abrahum/nonebot_plugin_r6s"
documentation = "https://github.com/abrahum/nonebot_plugin_r6s#readme"
